Crate libra_wallet

Re-exports

pub use crate::mnemonic::Mnemonic;

pub use crate::wallet_library::WalletLibrary;

Modules

error
Error crate

internal_macros
Internal macros The following macros are slightly modified from rust-bitcoin. The original file may be found here:

io_utils
Utils for read/write A module to generate, store and load known users accounts. The concept of known users can be helpful for testing to provide reproducible results.

key_factory
Utils for key derivation The following is a minimalist version of a hierarchical key derivation library for the LibraWallet.

mnemonic
Utils for mnemonic seed The following is a slightly modified version of the file with the same name in the rust-wallet crate. The original file may be found here:

wallet_library
Utils for wallet library The following document is a minimalist version of Libra Wallet. Note that this Wallet does not promote security as the mnemonic is stored in unencrypted form. In future iterations, we will be realesing more robust Wallet implementations. It is our intention to present a foundation that is simple to understand and incrementally improve the LibraWallet implementation and it's security guarantees throughout testnet. For a more robust wallet reference, the authors suggest to audit the file of the same name in the rust-wallet crate. That file can be found here: