Module libra_wallet::key_factory

Utils for key derivation The following is a minimalist version of a hierarchical key derivation library for the LibraWallet.

Note that the Libra Blockchain makes use of ed25519 Edwards Digital Signature Algorithm (EdDSA) and therefore, BIP32 Public Key derivation is not available without falling back to a non-deterministic Schnorr signature scheme. As LibraWallet is meant to be a minimalist reference implementation of a simple wallet, the following does not deviate from the ed25519 spec. In a future iteration of this wallet, we will also provide an implementation of a Schnorr variant over curve25519 and demonstrate our proposal for BIP32-like public key derivation.

Note further that the Key Derivation Function (KDF) chosen in the derivation of Child Private Keys adheres to HKDF RFC 5869.

Structs

ChildNumber
A child number for a derived key, used to derive a certain private key from the Master

ExtendedPrivKey
Derived private key.

KeyFactory
Wrapper struct from which we derive child keys

Master
Master is a set of raw bytes that are used for child key derivation

Seed
Seed is the output of a one-way function, which accepts a Mnemonic as input