Struct libra_wallet::WalletLibrary

pub struct WalletLibrary { /* fields omitted */ }

WalletLibrary contains all the information needed to recreate a particular wallet

Methods

impl WalletLibrary

pub fn new() -> Self

Constructor that generates a Mnemonic from OS randomness and subsequently instantiates an empty WalletLibrary from that Mnemonic

pub fn new_from_mnemonic(mnemonic: Mnemonic) -> Self

Constructor that instantiates a new WalletLibrary from Mnemonic

pub fn mnemonic(&self) -> String

Function that returns the string representation of the WalletLibrary Mnemonic NOTE: This is not secure, and in general the mnemonic should be stored in encrypted format

pub fn write_recovery(&self, output_file_path: &Path) -> Result<()>

Function that writes the wallet Mnemonic to file NOTE: This is not secure, and in general the Mnemonic would need to be decrypted before it can be written to file; otherwise the encrypted Mnemonic should be written to file

pub fn recover(input_file_path: &Path) -> Result<WalletLibrary>

Recover wallet from input_file_path

pub fn key_leaf(&self) -> u64

Get the current ChildNumber in u64 format

pub fn generate_addresses(&mut self, depth: u64) -> Result<()>

Function that iterates from the current key_leaf until the supplied depth

pub fn new_address_at_child_number(
&mut self,
child_number: ChildNumber
) -> Result<AccountAddress>

Function that allows to get the address of a particular key at a certain ChildNumber

pub fn new_address(&mut self) -> Result<(AuthenticationKey, ChildNumber)>

Function that generates a new key and adds it to the addr_map and subsequently returns the AuthenticationKey associated to the PrivateKey, along with it's ChildNumber

pub fn get_addresses(&self) -> Result<Vec<AccountAddress>>

Returns a list of all addresses controlled by this wallet that are currently held by the addr_map

pub fn sign_txn(&self, txn: RawTransaction) -> Result<SignedTransaction>

Simple public function that allows to sign a Libra RawTransaction with the PrivateKey associated to a particular AccountAddress. If the PrivateKey associated to an AccountAddress is not contained in the addr_map, then this function will return an Error